requirements.txt
absl-py==2.1.0
addict==2.4.0
annotated-types==0.7.0
cachetools==5.5.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Deprecated==1.2.15
einops==0.7.0
et-xmlfile==1.1.0
fonttools==4.47.0
google-auth==2.37.0
google-auth-oauthlib==1.0.0
grpcio==1.69.0
humanize==4.10.0
idna==3.6
imageio==2.35.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.4
mdurl==0.1.2
mmcv==1.3.8
mmsegmentation==0.14.1
networkx==3.1
nibabel==5.2.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
pip==23.3.1
platformdirs==4.1.0
prettytable==3.9.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-hostlist==1.23.0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
segm==0.0.1
setuptools==68.2.2
shellingham==1.5.4
SimpleITK==2.4.1
six==1.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tifffile==2023.7.10
timm==0.4.12
tomli==2.0.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchio==0.20.1
torchvision==0.12.0+cu113
tqdm==4.66.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
universeg==0.1.0
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.6
wheel==0.41.2
wrapt==1.17.2
yapf==0.40.2
zipp==3.17.0
